+++ js/advertForm/advertFormMessages.js
@@ -1,0 +1,7 @@
+const setInvalidMessageRoom = () => 'Количество комнат не подходит для количества гостей';
+const setInvalidMessageCapacity = () => 'Количество гостей не подходит для количества комнат';
+
+export {
+  setInvalidMessageRoom,
+  setInvalidMessageCapacity,
+};
