+++ js/advertForm/advertFormPristineConfig.js
@@ -1,0 +1,10 @@
+const configPristineAdvertFrom = () => ({
+  classTo: 'ad-form__element',
+  errorClass: 'ad-form__element--invalid',
+  successClass: 'ad-form__element--valid',
+  errorTextParent: 'ad-form__element',
+  errorTextTag: 'span',
+  errorTextClass: 'ad-form__message'
+});
+
+export {configPristineAdvertFrom};
