+++ js/advertForm/gettersAdvertForm.js
@@ -1,0 +1,3 @@
+const getValidStatus = (options, roomsComponent, capacityComponent) => options[roomsComponent.value].includes(Number(capacityComponent.value));
+
+export {getValidStatus};
