+++ js/main.js
@@ -1,10 +1,12 @@
 import {createTemplateCard} from './generationTemplate.js';
 import {createSimilarAdverts} from './data.js';
 import {getElementFromDocument} from './getterGenerationTemplate.js';
-import './advert-form.js';
+import {setValidateAdvertForm} from './advertForm/advertForm.js';
 
 
 const placeExample = getElementFromDocument('#map-canvas');
 const [{author, offer}] = createSimilarAdverts();
 
 placeExample.append(createTemplateCard(author, offer));
+
+setValidateAdvertForm();
